re PR testsuite/52641 (Test cases fail for 16-bit int targets)
	PR testsuite/52641
	* gcc.dg/misaligned-expand-1.c (cst): Cast to int.
	* gcc.dg/misaligned-expand-2.c (cst): Likewise.

From-SVN: r185602

+++ gcc/testsuite/gcc.dg/misaligned-expand-1.c
@@ -14,7 +14,7 @@ foo (myint *p)
   return *p;
 }
 
-#define cst 0xdeadbeef
+#define cst (int) 0xdeadbeef
 #define NUM 8
 
 struct blah
